requirements.txt
Django==5.1.7
asgiref==3.8.1
sqlparse==0.5.3
tzdata==2025.2
python-decouple==3.8
django-crispy-forms==2.1
Pillow==10.2.0
djangorestframework==3.14.0
psycopg2-binary==2.9.9
redis==5.0.1
django-redis==5.4.0
django-filter==23.5
djangorestframework-simplejwt==5.3.1
argon2-cffi==23.1.0
gunicorn==21.2.0
django-cors-headers==4.7.0